requirements.txt
openai==1.45.0
prompt_toolkit==3.0.47
python-dotenv==1.0.1
rich==13.8.1
anthropic==0.34.2
google-generativeai==0.8.1
